Refactor signup&in classes
